feat: add incentivize command
